Minimum necessary upgrades for solver scheduler to pass tests in current nova
This patch includes the minimum updates needed to fill the gap between solver
scheduler and current status of nova. It ensures that solver scheduler's tests
can pass under the latest nova environment. The patch does not include changes
that match all features in current nova codes. More changes are needed in the
future to fully port solver scheduler to the current nova version.

The main changes in this patch include:
- Switch from incubated depencies to corresponding oslo_* libraries.
- Minimum and necessary changes in solver_scheduler_host_manager,
solver_scheduler, and several unit testing modules to ensure that tests can
pass in the latest nova environment.
- Removed some unneeded tests which are not compatible with current nova
environment.

TODO:
Future changes should cover all the features/changes in current nova but not
in nova solver scheduler. Testings need to be enhanced to ensure the coverage
and compatibility to current and future nova versions.

Change-Id: Ib37b9361d3824fc12620db72892fcdcd0cab09af
Implements: blueprint update-to-liberty