Minimum necessary upgrades for solver scheduler to pass stable/juno tests
This patch includes the minimum updates needed to fill the gap between solver
scheduler stable/icehouse and stable/juno. It ensures that solver scheduler's
tests can pass under the juno version of nova environment. The patch does not
include changes that match all features in nova stable/juno. More changes are
needed in the future to fully port solver scheduler to the juno version.

The main changes in this patch include:
- Necessary changes in solver_scheduler
- Several unit testing modules to ensure that tests can pass in the juno
environment.
- Removed some unneeded tests which are not compatible with the juno version
environment.

Change-Id: Ibaa0ec2295232ffb7804817e563a0be4e430b59f
Implements: blueprint update-to-juno