Add rack affinity costs for network optimization
This patch adds 2 cost classes.
- TenantRackAffinityCost: place instances in the racks that contain existing
instances of the tenant. If a rack has existing instances of the same tenant
as that making request, then the hosts in the rack will get lower cost value.
- RackNetworkAffinityCost: place instances in the racks that contain existing
instances which are connected to the same tenant networks as requested
instances. If a rack has existing instances that are connected to the same
networks as those of requested instances, then the hosts in the rack will get
lower cost value.

Change-Id: I34cdeb6936957af7ee4c07d4f4fd0d91cf4a358f