Add tenant rack constraint to solver scheduler
This helps to limit the maximum number of racks that each tenant's VMs can
spread across. The max rack number can be set via config option, and if the
limit is not reached for a tenant the constraint logic will automatically look
for additional rack(s) and mark hosts in the additional rack(s) as available.

Change-Id: I792ac2e0ad75a6df8e8874b159a26760c47cfbd7